[Cases] Extract makeFileComment to a non-test module to fix flaky utils.test (#286099)
Fixes #286087 - likely introduced by #272759 (cc `@christineweng`)

### Summary

- `utils.test.tsx` imported `makeFileComment` from
`./case_view_files.test`, so loading it re-registered and re-ran the
**entire** `CaseViewFiles` RTL suite inside the `utils.test.tsx` worker.
- That overloaded worker pushed the `intersect with caseData.comments`
render test past Jest's 10s per-test budget under CI load — the reported
timeout.
- This patch moves `makeFileComment` into a plain (non-`.test`)
`test_helpers.ts` module and imports it from both
`case_view_files.test.tsx` and `utils.test.tsx`, so `utils.test.tsx` no
longer drags in the render suite.

### Context

- Follows the [Failed Test Investigator
analysis](https://github.com/elastic/kibana/issues/286087#issuecomment-5343987644)
(posted 2026-08-19): a `.test` import overloads the Jest worker; not a
product race (`useGetCaseFiles` is mocked, so `CaseViewFiles` renders
synchronously). I extracted only `makeFileComment` — the sole cross-file
helper — and left `makeFile` local to `case_view_files.test.tsx`, since
nothing else imports it.
- A single failure on `kibana-on-merge` / `main` ([build
#106921](https://buildkite.com/elastic/kibana-on-merge/builds/106921));
the issue's reported Location is `utils.test.tsx` while the timed-out
`it` lives in `case_view_files.test.tsx:144`, which is exactly the
cross-file-import fingerprint.

<details>
<summary>Verification</summary>

#### Verified locally

✅ Passed: `node scripts/eslint <changed files>`
✅ Passed: `node scripts/jest .../utils.test.tsx`: workload cut confirmed
— before the fix the run executed **187 tests in ~33.3s** and included
the `Case View Page files tab` suite; after the fix it runs **181 tests
in ~4.6s** and that suite is gone from this worker.
✅ Passed: `node scripts/jest .../case_view_files.test.tsx`: 6/6 tests
pass standalone (~6.3s for the whole file), so the render suite still
runs in its own worker.

This is a deterministic workload duplication, not a timing race, so a
flaky-repeat loop isn't the relevant signal — removing the duplicated
suite from the `utils` worker is the fix, shown directly by the
test-count and duration drop above.

#### Not verified locally

- Behavior under CI parallel load (the original timeout only surfaced
under contention) — cannot be reproduced on a single-worker local run.

</details>

<details>
<summary>Backporting guidance</summary>

Applied `backport:version` + `v9.5.2`. The bad `./case_view_files.test`
import and both test files exist on the `9.5` branch, so it is flaky
there too and the same extraction applies. `9.4` predates PR #272759
(its `utils.test.tsx` has no such import and `case_view_files.test.tsx`
does not exist there), and neither file exists on `8.19`, so those
branches need no fix.

</details>

> [!NOTE]
> Requested by @kibanamachine. Share feedback in #kibana-qa. Mention
`@copilot` to make quick changes.

> [!WARNING]
> <details>
> <summary>Firewall blocked 1 domain</summary>
>
> The following domain was blocked by the firewall during workflow
execution:
>
> - `awmgmcpg`
>> To allow these domains, add them to the `network.allowed` list in
your workflow frontmatter:
>
> ```yaml
> network:
>   allowed:
>     - defaults
>     - "awmgmcpg"
> ```
>
> See [Network
Configuration](https://github.github.com/gh-aw/reference/network/) for
more information.
>
> </details>

Fixer](https://github.com/elastic/kibana/actions/runs/32267695778) for
#286087 · 289.6 AIC · ⌖ 17.7 AIC · ⊞ 10.7K ·
[◷](https://github.com/search?q=repo%3Aelastic%2Fkibana+%22gh-aw-workflow-id%3A+flaky-test-fixer%22&type=pullrequests)

<!-- gh-aw-agentic-workflow: Flaky Test Fixer, engine: claude, version:
2.1.165, model: opus, id: 32267695778, workflow_id: flaky-test-fixer,
run: https://github.com/elastic/kibana/actions/runs/32267695778 -->

<!-- gh-aw-workflow-id: flaky-test-fixer -->
<!-- gh-aw-workflow-call-id: elastic/kibana/flaky-test-fixer -->

(cherry picked from commit 4564dc931f1b73e39e20ad56410b3ea258feb983)
